pyproject.toml
[tool.poetry]
name = "adjust-home-task"
version = "0.1.0"
description = ""
authors = ["Nikita Kolybelkin <kennedywhokilled@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
djangorestframework = "^3.13.1"
Django = "^4.0.2"
django-filter = "^21.1"

[tool.poetry.dev-dependencies]
pytest = "^7.0.0"
black = "^22.1.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
